% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_renormalise.R
\name{compute_fraction_from_count}
\alias{compute_fraction_from_count}
\title{Convert counts into fractions of total counts}
\usage{
compute_fraction_from_count(counts)
}
\arguments{
\item{counts}{a numeric vector of counts}
}
\value{
a numeric vector of fractions
}
\description{
Convert counts into fractions of total counts, dealing with zero sum counts by setting to zero instead of Inf
}
\examples{
compute_fraction_from_count(c(10, 2, 1000, 50, 250))

}
