% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_maths.R
\name{sig_add}
\alias{sig_add}
\alias{\%+\%}
\title{Add Two Catalogues}
\usage{
sig_add(catalogue1, catalogue2)

catalogue1 \%+\% catalogue2
}
\arguments{
\item{catalogue1, catalogue2}{Two \code{sigverse} catalogue data.frames.
See \code{\link[sigshared:catalogue]{sigshared::example_catalogue()}}.
Must have the same \code{type} and \code{channel} rows in the same order.}
}
\value{
A single \code{sigverse} catalogue data.frame with summed \code{count}
and recomputed \code{fraction} values. See \code{\link[sigshared:catalogue]{sigshared::example_catalogue()}}.
}
\description{
Sums two sigverse-style catalogues element-wise by their \code{count} values.
}
\details{
This operator provides a concise way to add two catalogues. For summing multiple
catalogues, use \code{\link[=sig_sum]{sig_sum()}}.
}
\examples{
library(sigstash)
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")
cat1 <- sig_reconstruct(signatures[['SBS3']], n = 100)
cat2 <- sig_reconstruct(signatures[['SBS4']], n = 100)

# Add two catalogues
cat_sum <- cat1 \%+\% cat2

# Add multiple catalogues via repeated \%+\%
cat3 <- sig_reconstruct(signatures[['SBS5']], n = 100)
total <- cat1 \%+\% cat2 \%+\% cat3

# Addition two signatures using the full function name
sig_add(cat1, cat2)

# Alternatively, use sig_sum for a list of catalogues
catalogues <- list(cat1 = cat1, cat2 = cat2, cat3 = cat3)
total2 <- sig_sum(catalogues)
}
\seealso{
\code{\link[=sig_sum]{sig_sum()}} for summing a collection of catalogues
}
