% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_maths.R
\name{sig_combine}
\alias{sig_combine}
\title{Add Signatures to a Combined Signature Model}
\usage{
sig_combine(signatures, model)
}
\arguments{
\item{signatures}{A signature collection, typically a list of data.frames where each data.frame represents a signature with specific attributes.}

\item{model}{A numeric vector representing the contribution of each signature to the combined model. The sum of the values in this vector should be less than or equal to 1.}
}
\value{
A data.frame in the 'combined_signature_model' style containing the selected signatures and their modified fractions based on the model.
}
\description{
This function takes a signature collection and a model as input and adds selected signatures to a combined signature model. The output is a data.frame in the 'combined_signature_model' style data.frame format from the Sigverse package.
}
\examples{
library(sigstash)

# Load a signature collection
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Create a model that represents a mix of SBS1 (40\%) and SBS2 (60\%)
model <- c(SBS1 = 0.4, SBS2 = 0.6)

# Add selected signatures to the combined model
combined_signatures <- sig_combine(signatures, model)
print(combined_signatures)

# Visualise using sigvis

}
\seealso{
\code{\link{assert_signature_collection}}, \code{\link{assert_numeric_vector}}, \code{\link{assert_subset}}
}
