% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_renormalise.R
\name{sig_opportunities}
\alias{sig_opportunities}
\title{Autosomal context counts}
\usage{
sig_opportunities()
}
\value{
list of context counts generated by \href{https://github.com/CCICB/contextcounter}{contextcounter}
}
\description{
Autosomal context counts
}
\examples{
sig_opportunities()

}
