% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_renormalise.R
\name{sig_renormalise}
\alias{sig_renormalise}
\title{Signature renormalisation}
\usage{
sig_renormalise(signature, opportunities_original, opportunities_new)
}
\arguments{
\item{signature}{sigverse style signature data.frames. See \code{\link[sigshared:signature]{sigshared::example_signature()}}.}

\item{opportunities_original}{named vector of mutation opportunities for each context. See \code{\link[=sig_opportunities]{sig_opportunities()}}.}

\item{opportunities_new}{named vector of mutation opportunities for each context. See \code{\link[=sig_opportunities]{sig_opportunities()}}.}
}
\value{
sigverse style signature data.frame. See \code{\link[sigshared:signature]{sigshared::example_signature()}}.
}
\description{
Transform a signature into what it would look like if called from a different pipeline/genome/organism to account for the different 'mutation opportunites'
}
\examples{
opps <- sig_opportunities()
signature <- data.frame(
  channel = c("A[C>G]G", "A[C>G]C", "A[C>G]T"),
  type = c("C>G", "C>G", "C>G"),
  fraction = c(0.4, 0.1, 0.5)
)

# Convert genome signature to panel signature (TSO 500)
sig_renormalise(
  signature,
  opps$GRCh38_genome_trinucleotide_autosomal,
  opps$GRCh38_tso500_trincucleotide_autosomal
)
}
