% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_maths.R
\name{sig_sum}
\alias{sig_sum}
\title{Sum a Collection Catalogues}
\usage{
sig_sum(catalogues)
}
\arguments{
\item{catalogues}{A named list of \code{sigverse} catalogue data.frames. See \code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}}
}
\value{
A \code{sigverse} catalogue data.frame representing the total. See \code{\link[sigshared:catalogue]{sigshared::example_catalogue()}}.
}
\description{
Sums a list of sigverse catalogues, or two individual catalogues, into a single result.
}
\details{
This function is useful for aggregating catalogues across samples or replicates.
If you only need to add two catalogues, you may use the \code{\link{\%+\%}} operator instead.
}
\examples{
library(sigstash)

# Load a signature collection
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Reconstruct catalogues for two pure samples (each with 100 mutations)
catalogue1 <- sig_reconstruct(signatures[['SBS3']], n = 100)
catalogue2 <- sig_reconstruct(signatures[['SBS4']], n = 100)

# Sum catalogue1 and  catalogue2
catalogue_sum <- catalogue1 \%+\% catalogue2

# Sum a collection
collection <- list(cat1 = catalogue1, cat2 = catalogue2)
collection <- sig_sum(collection)
}
\seealso{
\code{\link{\%+\%}} for summing two catalogues
}
