% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{sig_summarise_bootstraps}
\alias{sig_summarise_bootstraps}
\title{Summarise Signature Bootstraps}
\usage{
sig_summarise_bootstraps(
  bootstraps,
  threshold = 0.05,
  outliers_as_strings = FALSE
)
}
\arguments{
\item{bootstraps}{A data frame containing bootstrap results in a sigverse-style format.
See \code{\link[sigshared:bootstraps]{sigshared::example_bootstraps()}} for details.}

\item{threshold}{The minimum proportion of mutations that must be explained by a signature for it to be
considered present in a bootstrap. This threshold is used for calculating
the experimental p-value via \code{\link[=sig_compute_experimental_p_value]{sig_compute_experimental_p_value()}}.}

\item{outliers_as_strings}{Logical. Return outliers as a character vector of '|' delimited strings rather than as a list-column (ignored if \code{return_dataframe = FALSE}).
Can customise the delimiter using the \code{delim} argument}
}
\value{
A data frame containing the following columns:
\item{signatures}{Names of the signatures.}
\item{quantiles, iqr, min, max,outlier_thresholds}{Statistical summary columns (quantiles, min, and max) of the contribution proportions across bootstraps.}
\item{outliers}{either a list column or a character of '|' outlier strings (depending on \code{outliers_as_strings} argument}
\item{p_value}{Computed experimental p-value for each signature, based on the threshold.}
}
\description{
This function computes summary statistics for signature-level bootstraps,
including quantiles, min/max contribution proportions, and an experimental p-value.
}
\examples{
library(sigshared)
bootstraps <- example_bootstraps()
sig_summarise_bootstraps(bootstraps, threshold = 0.05)
}
