% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_maths.R
\name{sig_combine}
\alias{sig_combine}
\title{Add Signatures to a Combined Signature Model}
\usage{
sig_combine(
  signatures,
  model,
  format = c("signature", "combined"),
  verbose = FALSE
)
}
\arguments{
\item{signatures}{A sigverse signature collection. See (\code{\link[sigshared:signature_collection]{sigshared::example_signature_collection()}}) for details.}

\item{model}{A named numeric vector representing the contribution of each signature to the combined model.
The names correspond to the signatures, and the values represent their contributions.
The sum of the values in this vector should be less than or equal to 1.}

\item{format}{A character string indicating the output format.
If "combined", the function returns a 'combined_signature_model' data.frame where each row represents a contribution for a particular channel from a single signature (duplicate channels are not collapsed).
If "signature", the function returns the data in the sigverse signature format, representing a novel signature created by combining the signatures in the collection according to the ratios described by the model.}

\item{verbose}{enables detailed output messages (flag).}
}
\value{
A data.frame in the 'combined_signature_model' style containing the selected signatures and their modified fractions based on the model.
}
\description{
This function takes a signature collection and a model as input and adds
selected signatures to a combined signature model.
The default output is a data.frame in either sigverse signature format or the 'combined_signature_model'
style depending on \code{format}.
}
\examples{
library(sigstash)

# Load a signature collection
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Create a model that represents a mix of SBS1 (40\%) and SBS2 (60\%)
model <- c(SBS1 = 0.4, SBS2 = 0.6)

# Add selected signatures to the combined model
combined_signatures <- sig_combine(signatures, model)
print(combined_signatures)


}
\seealso{
\code{\link{example_signature_collection}} \code{\link{example_model}}
}
