% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_maths.R
\name{sig_subtract}
\alias{sig_subtract}
\title{Subtract Signature}
\usage{
sig_subtract(signature1, signature2)
}
\arguments{
\item{signature1, signature2}{sigverse signature data.frames}
}
\value{
a data.frame representing a sigverse signature
}
\description{
Subtracts signature2 from signature1 (factions) and returns result.
Works for only signatures. To translate to a catalogue see \code{\link[=sig_reconstruct]{sig_reconstruct()}}
}
\examples{
library(sigstash)

# Load a signature collection
signatures <- sig_load("COSMIC_v3.3.1_SBS_GRCh38")

# Subtract signatures
sig_subtract(signatures[['SBS3']], signatures[['SBS4']])

}
